HG-171 Backend Error Messages

+++ backend-hedera/custodian/apps/custodian/src/user/controller/user.controller.ts
@@ -8,6 +8,8 @@ import {
     Put,
     Delete,
     Query,
+    HttpException,
+    HttpStatus,
 } from '@nestjs/common';
 import { UsersDTO } from '@app/shared/users/dto/users.dto';
 import { AuthGuardService } from '@app/core/auth-guard/service/auth-guard.service';
@@ -17,6 +19,7 @@ import { UserUpdateDto } from '@app/shared/users/dto/user-update.dto';
 import { DataExportQueryDto } from '@app/shared/util/dto/data.export.query.dto';
 import { UserStateConstant } from '@app/shared/users/constants/user.state.constants';
 import { UserService } from '@app/shared/users/service/user.service';
+import { RoleEnum } from '@app/shared/role/enum/role.enum';
 
 @Controller('user')
 export class UserController {
@@ -25,6 +28,18 @@ export class UserController {
     @UseGuards(AuthGuardService)
     @Post('add')
     async add(@Body() userDto: UsersDTO, @Request() req): Promise<any> {
+        if (
+            req?.user &&
+            (req?.user?.userRole !== RoleEnum.Admin ||
+                req?.user?.userRole !== RoleEnum.Root)
+        ) {
+            throw new HttpException(
+                userDto.company
+                    ? 'You do not have permission to create a new organisation.'
+                    : 'You do not have permission to create a new user.',
+                HttpStatus.UNAUTHORIZED,
+            );
+        }
         return this.userService.register(
             userDto,
             '',
@@ -64,6 +79,15 @@ export class UserController {
     @UseGuards(AuthGuardService)
     @Put('update')
     updateUser(@Body() userUpdate: UserUpdateDto, @Request() req) {
+        if (
+            req?.user &&
+            (req?.user?.userRole !== RoleEnum.Admin ||
+                req?.user?.userRole !== RoleEnum.Root)
+        )
+            throw new HttpException(
+                'You do not have permission to delete users.',
+                HttpStatus.UNAUTHORIZED,
+            );
         return this.userService.updateUserDetails(userUpdate, req.user);
     }
 
@@ -76,6 +100,15 @@ export class UserController {
     @UseGuards(AuthGuardService)
     @Delete('delete')
     deleteUser(@Query('userId') userId: number, @Request() req) {
+        if (
+            req?.user &&
+            (req?.user?.userRole !== RoleEnum.Admin ||
+                req?.user?.userRole !== RoleEnum.Root)
+        )
+            throw new HttpException(
+                'You do not have permission to delete users.',
+                HttpStatus.UNAUTHORIZED,
+            );
         return this.userService.deleteUser(userId, req.user);
     }
 }
